+++ module-apps/application-settings-new/ApplicationSettings.hpp
@@ -64,6 +64,8 @@ namespace app
             virtual ~OperatorsSettings()                               = default;
             virtual void setOperatorsOn(bool value)                    = 0;
             [[nodiscard]] virtual bool getOperatorsOn() const noexcept = 0;
+            virtual void setVoLTEOn(bool value)                        = 0;
+            [[nodiscard]] virtual bool getVoLTEOn() const noexcept     = 0;
         };
     }; // namespace settingsInterface
 
@@ -93,12 +95,16 @@ namespace app
         void operatorOnChanged(const std::string &value);
         void setOperatorsOn(bool value) override;
         bool getOperatorsOn() const noexcept override;
+        void setVoLTEOn(bool value) override;
+        bool getVoLTEOn() const noexcept override;
+        void volteChanged(const std::string &value);
 
       private:
         Store::GSM::SIM selectedSim   = Store::GSM::get()->selected;
         std::string selectedSimNumber = {};
         bsp::Board board              = bsp::Board::none;
         bool operatorsOn              = false;
+        bool voLteStateOn             = false;
     };
 
     template <> struct ManifestTraits<ApplicationSettingsNew>